refactor: component data structure and improve component categories section in home page
