+++ lib/data/components/application/index.ts
@@ -1,0 +1,20 @@
+import { ComponentType } from '../types';
+import { buttons } from './buttons';
+import { cards } from './cards';
+import { alerts } from './alerts';
+import { forms } from './forms';
+import { badges } from './badges';
+import { avatars } from './avatars';
+import { progress } from './progress';
+import { skeleton } from './skeleton';
+
+export const applicationComponents: ComponentType[] = [
+  buttons,
+  cards,
+  alerts,
+  forms,
+  badges,
+  avatars,
+  progress,
+  skeleton,
+];
