+++ lib/data/components/marketing/index.ts
@@ -1,0 +1,16 @@
+import { ComponentType } from '../types';
+import { heroes } from './heroes';
+import { ctas } from './ctas';
+import { footers } from './footers';
+import { pricing } from './pricing';
+import { testimonials } from './testimonials';
+import { faq } from './faq';
+
+export const marketingComponents: ComponentType[] = [
+  heroes,
+  ctas,
+  footers,
+  pricing,
+  testimonials,
+  faq,
+];
